工数入力の時間が浮動小数点でなかった。

+++ src/Nulab.Backlog.Api/Data/Responses/Issue.cs
@@ -36,9 +36,9 @@
 
         public DateTime? dueDate { get; set; }
 
-        public int? estimatedHours { get; set; }
+        public double? estimatedHours { get; set; }
 
-        public int? actualHours { get; set; }
+        public double? actualHours { get; set; }
 
         public int? parentIssueId { get; set; }
